fixed consumer deserializer
