--- domain/src/main/java/co/crisi/shipm8/domain/event/FailedOrderDto.java
@@ -1,5 +1,0 @@
-package co.crisi.shipm8.domain.event;
-
-public record FailedOrderDto(Long orderId, String errorMessage) {
-
-}
