+++ domain/src/main/java/co/crisi/shipm8/domain/message/FailedOrderDto.java
@@ -1,0 +1,7 @@
+package co.crisi.shipm8.domain.message;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record FailedOrderDto(@JsonProperty("orderId") Long orderId, @JsonProperty("error") String errorMessage) {
+
+}
