+++ domain/src/main/java/co/crisi/shipm8/domain/message/OrderDto.java
@@ -1,4 +1,4 @@
-package co.crisi.shipm8.domain.event;
+package co.crisi.shipm8.domain.message;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
