+++ domain/src/main/java/co/crisi/shipm8/domain/message/OrderProcessed.java
@@ -1,4 +1,4 @@
-package co.crisi.shipm8.domain.event;
+package co.crisi.shipm8.domain.message;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
