+++ domain/src/main/java/co/crisi/shipm8/service/OrderServicePort.java
@@ -4,9 +4,9 @@ import co.crisi.shipm8.domain.IOrder;
 import co.crisi.shipm8.domain.IProduct;
 import co.crisi.shipm8.domain.data.Order;
 import co.crisi.shipm8.domain.data.input.OrderSaveDto;
-import co.crisi.shipm8.domain.event.OrderDto;
-import co.crisi.shipm8.domain.event.OrderProcessed;
-import co.crisi.shipm8.domain.event.ProductUpdateDto;
+import co.crisi.shipm8.domain.message.OrderDto;
+import co.crisi.shipm8.domain.message.OrderProcessed;
+import co.crisi.shipm8.domain.message.ProductUpdateDto;
 import co.crisi.shipm8.exception.business.AddressNotFoundException;
 import co.crisi.shipm8.exception.business.BusinessException;
 import co.crisi.shipm8.exception.business.OrderNotFoundException;
